refactor files (work in progress)
